moved CoursePane to RightPane/CoursePane
