+++ src/components/RightPane/AddedCourses/CustomEventDetailView.js
@@ -2,10 +2,10 @@ import React from 'react';
 import { withStyles } from '@material-ui/core/styles';
 import { Card, CardActions, CardHeader, IconButton } from '@material-ui/core';
 import { Delete } from '@material-ui/icons';
-import ColorPicker from '../ColorPicker';
+import ColorPicker from '../../ColorPicker';
 import moment from 'moment';
-import { deleteCustomEvent } from '../../actions/AppStoreActions';
-import CustomEventDialog from '../Calendar/CustomEvents/CustomEventDialog';
+import { deleteCustomEvent } from '../../../actions/AppStoreActions';
+import CustomEventDialog from '../../Calendar/CustomEvents/CustomEventDialog';
 import ReactGA from 'react-ga';
 
 const styles = {
